@@ -43,9 +43,14 @@ class UserService extends BaseService
         }
 
         return DB::transaction(function () use ($id, $data) {
-            $user = $id ? $this->repository->update($id, $data) : $this->repository->create($data);
+            $user = $id
+                ? $this->repository->update($id, $data)
+                : $this->repository->create($data);
+
+            // Roles
             $user->syncRoles($data['role_ids']);
 
+            // File
             $this->fileService->storeFile($user, $data);
 
             return $user;
